+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-08-05
+
+### Added
+
+- 1.19.4 support
+- Forge support
+- Occultism integration
+- IntegratedDynamics integration
+- Nature's Compass integration
+- Easy Villager integration
+- Tom's Simple Storage integration
+
 ## [0.2.4] - 2023-04-27
 
 ### Fixed